requirements.txt
python-telegram-bot[job-queue]==22.7
psycopg2-binary
stripe